fix(optimization): extract shared fingerprinting concern and cap surrogate model rows
Extract canonicalize, custom_prompt_sha256, normalized_service_container_ids,
and normalized_mcp_servers into ConfigurationBundles::BundleFingerprinting
to eliminate duplication between AssignToRun and Optimizer that could cause
fingerprint divergence.

Cap SurrogateModel outcome loading to 500 most recent rows to prevent
unbounded memory growth as projects accumulate runs.
