+++ app/services/configuration_bundles/assign_to_run.rb
@@ -4,6 +4,8 @@ require "digest"
 
 module ConfigurationBundles
   class AssignToRun
+    include BundleFingerprinting
+
     attr_reader :agent_run
 
     def initialize(agent_run:)
@@ -50,22 +52,6 @@ module ConfigurationBundles
       )
     end
 
-    def custom_prompt_sha256
-      return if agent_run.custom_prompt.blank?
-
-      Digest::SHA256.hexdigest(agent_run.custom_prompt)
-    end
-
-    def normalized_service_container_ids
-      ids = Array(agent_run.service_container_ids).map { |id| Integer(id, exception: false) || id }.compact.sort
-      ids if ids.any?
-    end
-
-    def normalized_mcp_servers
-      servers = Array(agent_run.mcp_server_snapshot).filter_map { |snapshot| snapshot["name"].presence }.sort
-      servers if servers.any?
-    end
-
     def experiment_definitions(selected_variants = nil)
       ConfigurationExperiment::TRACKED_CONFIG_KEYS.each_with_object({}) do |config_key, definitions|
         experiment = ConfigurationExperiment.active_for(config_key, project: agent_run.project, agent_run: agent_run)
@@ -81,19 +67,6 @@ module ConfigurationBundles
           configuration_experiment_variant_id: assignment.configuration_experiment_variant_id,
           value: assignment.configuration_experiment_variant.parsed_value
         }
-      end
-    end
-
-    def canonicalize(value)
-      case value
-      when Hash
-        value.each_with_object({}) do |(key, nested_value), normalized|
-          normalized[key.to_s] = canonicalize(nested_value)
-        end.sort.to_h
-      when Array
-        value.map { |nested_value| canonicalize(nested_value) }
-      else
-        value
       end
     end
